requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudinary==1.23.0
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.2.1
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
flake8==3.8.4
Flask==1.1.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pyparsing==2.4.6
python-dotenv==0.15.0
python-telegram==0.12.0
python-telegram-bot==13.0
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.20
tornado==6.1
tzlocal==2.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
